Rewrote USB CDC host driver to match design of other custom drivers
This rewrite is still substantially based on the code of the original
implementation, so it retains the same copyright notices. However, the
guts have changed significantly.  It now more closely resembles the
other custom USB serial drivers, maybe most closely the PLCOM one.

The change was made to more easily fit within the rest of the system,
where the drivers are expected to poll on their own with a regular
interval and invoke a callback when new data is available.
It was also made to add the ability to set line state flags.

In its current form, it depends on the same interrupt handler tweak
that's necessary for the PLCOM driver.  Its still unclear if that's the
right way to do things, but it seems to work well enough for now.