Add parsing for Tobias TBX2000C output format and fallback
This adds parsing for the data format seen out the USB port of the
Tobias TBX2000C transmission densitometer. Since its unclear how this
differs from other Tobias models, or even Tobias reflection models, the
reflection-vs-transmission flag is not set.

Furthermore, as a way of future-proofing the parser code, a fallback
parser has been added that simply attempts to parse the first number
sent to the device as if it were a visual density reading.