Profile name refactoring
Added ability to load enlarger profile names, separate from whole
enlarger profiles, so that the enlarger profile list is now far more
efficient to populate. This was done in preparation for a large
expansion in the size of an individual enlarger profile.

Added a constant for the length of profile names, in general, to replace
the need to hard-code "32" in far to many places within the code.
Given that this needs to be defined outside of the settings code, to
avoid circular dependencies, the settings code now uses compiler macros
to make certain the value is not inadvertently changed. This is because
any change to the value will break existing settings data.