Integrate DMX-capable enlarger control with the exposure timer
This integration is a bit invasive and violates some of the usual
abstractions, in the interests of keeping very consistent enlarger
timing.

During exposure mode, the DMX task is transitioned to the IDLE state and
the responsibility for triggering the actual sending of individual DMX
frames is given to the timer ISR inside the exposure timer module.
When in this state, the exposure timer ISR is also permitted to directly
update the raw DMX frame array since it knows when that array will and
won't be read from.

Since the exposure timer runs on a 10ms interval, and the standard DMX
frame takes 22.7ms to send, frames during exposure are sent at a 30ms
interval.  Ideally this should be reduced to 25ms, but doing that would
require changing the exposure timer to run on a 5ms interval.