Fix handling of blackout state on startup
When initializing the keypad, the initial state of the keys is now read
prior to enabling interrupt processing.

When the blackout handler callback is assigned, it is initially called
with the current state of the blackout switch. This makes it possible to
handle the current blackout state on device startup.

Only remaining glitch is that the device illumination will briefly flash
during startup, due to the order in which various components are
initialized. It may make sense to fix this in the future.