Refactor keypad controller I2C operations
There were many places where we were read/writing 3 adjacent registers
and handling the results in the same way.

First, this adds common functions to handle this operation, so the code
isn't duplicated for each register group.
Second, this switches to doing a single 3-byte read instead of three
separate 1-byte reads, for the sake of efficiency.