Move contrast filter settings from paper profile to enlarger config
Now that our enlarger configuration contains far more than a timing
profile, it makes sense to move the contrast filter setting here. Its
more of a property of the enlarger itself, and the paper profile is
really pinned to the totality of the darkroom setup.

Since this particular setting only applies to traditional color
enlargers, its hidden on DMX-controlled LED enlargers that are
unlikely to have dichroic filters.