Reworked enlarger calibration process for new sensor
Timing is now based on the tick time when the sensor interrupt is
received, rather than the timing of any sort of external sampling loop.

Results are now compiled using a geometric mean, for less sensitivity to
spurious readings.

There was an attempt to use the sensor in single-shot mode, for which
some additional functions and comments have been added. However, it was
determined that this mode has fixed overhead that makes it unusable for
this purpose.