Fix sensor scaling and add utility functions
The scaling of the sensor value is supposed to be 2^ALS_SCALE, but the
code was actually doing ALS_SCALE^2.  This is now fixed.

Also added utility functions to wrap the application of the scaling and
the conversion to basic counts, in addition to exposing the underlying
raw unscaled value.  This will be useful for judging signal level when
doing various calibration routines.