Add support for coalescing encoder events
The input encoder is capable of allowing the user to rapidly turn it to
rapidly adjust a value.  However, this sometimes means they can tick it
faster than the code on the other side can keep up (such as if its doing
display updates). This creates a lagged response, and may even risk key
event queue overflow.

The solution taken here is to update the keypad handler to use a queue
set for app-facing events. This set contains a normal queue for normal
key events, and a special single-item queue for encoder events.  As new
encoder events come in, that single item queue is updated with a count.

When key events are returned to the application, encoder events are
accompanied by this count. That way multiple subsequent ticks can be
processed as a single event with an increment value.

So far, areas of the application that directly handle key events and use
the encoder have been updated for this scheme.