Fix display assets so they stay only in flash
This frees up a little more RAM, though we really have plenty of that.
However, it should be noted that the best way of keeping track of unused
FLASH space is monitoring the size of the ".end_fill" ELF section.
That's not made completely obvious by the current build process.