Mark test strip exposure times that are too short for the enlarger
It may make sense to add some behavior here, since it is impossible to
create a correct test strip as-is when in this state.  Either prevent
test strip creation entirely, or come up with some sort of scheme to
only allow portions of a test strip that will get correct exposure.

Refs #25