Add basic I2C reinitialization on meter probe startup
Sometimes when hot-plugging the meter probe, the I2C bus may get stuck
in a busy state. If this is detected on the first I2C operation to the
probe, then reinitialize the I2C peripheral and try again.

Eventually a more complicated retry mechanism should probably be
implemented, including power toggling, but for now this seems to get
past the most common glitch.