Add ability to save/load meter probe calibration from a USB device
This feature is hopefully something that'll be seldom used by actual
users, and perhaps should even be disabled in the default firmware to
prevent accidents. However, extra confirmation has been added just in
case.

The idea is to make it possible to actually load calibration data into
the meter probe without the need for external hardware, even if the
process of actually creating that calibration data is less than
straightforward.

For real factory calibration of the meter probes, it is likely that an
external hardware rig will be constructed to connect to both the meter
probe and a lux meter, and to do all of the steps that way.