Moved footswitch connector and cleaned up 5V routing
Moved the footswitch connector back to give extra clearance for
enclosure jacks in that area.

Rearranged the footswitch circuitry to cluster it better and provide the
necessary space.

Rerouted the 5V supply trace to avoid having to make loops around the
board to go to the various front panel support circuits.

Refs #38