Initial implementation of light sensor oscillator calibration
Added the ability to calibrate the oscillator in the TSL2585 sensor at
startup. This works by configuring the interrupt pin to act as VSYNC
input on the sensor side, then routing TIM3/CH4 to that pin on the MCU
side to supply a 1kHz reference signal.

While this whole process involves a lot of reconfiguration of the sensor
and MCU, it runs rather quickly.  Whether it helps or hurts will require
further testing, so right now it is only enabled in diagnostic mode.