First pass at putting together a basic CMake based build process
Right now this is mostly within a single configuration for the base
project, and does not yet include the bootloader code. Eventually common
elements could be refactored out.

The idea is to make it possible to build the firmware outside of the
Eclipse CDT managed build process, which has benefits for build servers,
testing, and easier IDE flexibility.

Also finally wrote a script to extract binary size and free space
information out of the build output. This requires using the GNU
binutils "size" command, and parsing various bits of information from it
output.  The complexity here is unfortunately necessary, because our
binaries are intentionally padded out to the full flash size to support
checksums.

Some IDE config files for CLion have also been added, but setup
instructions are not yet written.