Add ability to handle missed sensor interrupts
If we're not able to handle an ALS interrupt in time, which most
commonly occurs when the integration cycle is either too fast or the MCU
is stopped in the debugger, the sensor's FIFO can fill up.  As the code
was previously implemented, this would put us into a wedged state.

The fix here is twofold.  First, if the FIFO contains more than one
reading, it is simply read in a loop until the last reading is captured
and returned.  Second, if the FIFO has overflowed, it is simply cleared
and an invalid reading is returned.