Make sure keypad controller FIFO is actually disabled in bootloader
This fixes a bug where the bootloader was enabling the keypad IC's FIFO
before the main firmware started up. On older hardware revisions, this
did not cause any problems. However, due to the default state of certain
keys changing in the Rev C hardware, it was causing unexpected events to
slip through into the startup process.