Keypad handling updates for Rev C hardware
The latest hardware revision changes some aspects of how the footswitch
and meter probe key buttons work, as followes:
- Both buttons are now active-high, rather than active-low
- The meter probe button will read high when the meter probe power is
  disabled.

The first step to resolving this is to reconfigure the keypad IC so that
both buttons now trigger on the opposite state from the rest of the
keys.

The second step is to make sure that all handling of the meter probe
button is completely disabled, unless the meter probe itself is actually
powered up.