Implemented updated meter target calibration
This assumes an out-of-band process has been performed to establish the
relationship between a calibrated lux measurement and a gain/slope
corrected meter reading.  Since this relationship has been found to be
linear, it is implemented as stored slope/intercept values based on a
linear regression of that process's data.