Switch sensor osc calibration to use SW_VSYNC_TRIGGER
The first attempt at oscillator calibration depended on reversing the
direction of the meter probe interrupt signal and making it a 1kHz timer
output. This somewhat worked, but was risky and occasionally
inconsistent because the sensor's logic-low voltage level was way too
close to the static offset induced by the buffers used to get the
signals into the meter probe. While there are plans to fix this in the
next revision of the meter probe, those changes will only cover the I2C
signals and not the interrupt pin.

The TSL2585 sensor provides another method of calibration, which can be
done entirely over I2C. It requires sending two consecutive I2C writes
with a known time interval between them. This change implements that
approach, using a timer interrupt to directly trigger those writes. It
seems to work with more consistent results than the previous approach.

This is still only used on the diagnostic screen, but it may be used for
normal operation in the future.