Finished CMake structure conversion for external modules
Everything in the "external" directory now has its own CMakeLists file,
and is built as a subdirectory. Dependencies are specified individually,
within those files.

Top-level includes still need to be propagated manually, because they
are where certain widely-reference config headers are located. That may
be worth fixing in the future.