Remove jump-to-system-memory feature from bootloader
This jumps to a built-in bootloader within the MCU, that is really not
useful to an end-user.  At the point where its desired to use the system
bootloader, one might as well already be opening the device to connect
a probe to the SWD port.

Removing this feature also simplifies the process of getting into the
normal USB-stick-load bootloader.