Initial implementation of FT260 in meter probe schematic
This is a big change towards using a USB interface for the meter probe,
and is the first step of getting the schematic portion put together.
Most of the designators will likely change after the layout is started.

The intent of this change is to leverage off-the-shelf mass produced mouse
cables for the physical connection, and to run the meter probe over a
protocol better suited to hot-plugging and noise immunity.