Initial schematic update for F411 to F446 switch
This includes the updated CubeMX configuration file, as well as the main
MCU symbol change on the main board schematic. No other changes have yet
been made, so this commit should be considered a work in progress.

Reasonable effort has been made to make as few pin changes as possible,
as both chips have the exact same package and are generally footprint
compatible, but some changes were necessary.

The driving motivation for this change is to gain access to a better USB
peripheral that has full host support by the CherryUSB library. Some
additional benefits include more timer capabilities and a higher system
clock speed.

The previous CubeMX files were moved into a subdirectory for easier
reference across the change, but may be removed at some point in the
future.

Refs #50