Finished migrating bootloader changes and it compiles
Code not yet tested.  The goal at this stage is to get enough of the
bootloader running to be able to launch the real firmware.
To actually do the real job of loading firmware into flash, the entire
USB layer of the bootloader will need to be changed to use a stack that
supports hubs.