Very basic initial FATFS integration with CherryUSB
This feels sort-of like a kludge to make the existing ST middleware fork
of FATFS integrate with our new CherryUSB stack.

The base FATFS code is quite outdated, and the integration lacks support
for multiple MSC devices. Both these things need to be addressed, but
the current solution is good enough for the time being.