Add a series of delays to the enumeration process to fix a strange bug
On some devices, transaction errors occur during the enumeration
process. Adding even the smallest delay after each transation during
this process seems to fix the issue.

While the root cause is still as of yet unknown, and this isn't the
greatest of fixes, this kludge is still just enough to get things moving
again to allow development progress to continue.