Add ability to use more complex logic for class driver selection
For many USB serial adapters, there may be multiple PIDs sharing the
same driver. The current design of CherryUSB class driver selection
makes this a difficult case to handle.

To improve the situation, the ability to provide a matching function as
part of the class driver interface has been added. This is hopefully a
simple patch to the upstream code, which won't be too difficult to
maintain going forward.