Created an initial fork of the CherryUSB FTDI driver
The FTDI driver needs the ability to handle a variety of device PIDs,
and a variety of chip versions. Furthermore, the baudrate selection code
depends on chip specifics.

To make it easier to integrate all of these changes, the upstream driver
has been forked (with a slight name change) into our local code base.

It is expected that similar things will need to be done with many of the
class drivers.