Disable protocol matching in CherryUSB CDC-ACM driver
Matching on the AT protocol is sort of a mistake, and one that the ST
stack also did by default. It makes it impossible to seemlessly support
all of the CDC devices that are not modems.

For now, making this trivial change against the library is the easiest
fix.  The class driver will only be forked if additional changes are
deemed necessary.