Consolidated FATFS integration code and added multiple MSC support
Moved all the bits of FATFS integration code into the same module as the
rest of the app-side USB stack MSC device handling.

Implemented support for multiple MSC devices, using the "lun" parameter
as a way of distinguishing them.

Still need to update the file picker to be aware of more than one
device.