Integrate FreeRTOS into the bootloader
This pulls in the FreeRTOS bits from the main firmware, and integrates
it into the bootloader. It is just the first step, but it is necessary
because the CherryUSB stack doesn't seem to support bare metal
operation.

The new bootloader startup flow will begin doing all of its startup
checks in "bare metal mode", then start the FreeRTOS scheduler only if
it actually enters bootloader mode to prompt for a firmware update.

More work will be necessary on the other side, to handle post-update
restart and handle some operations differently.  Also, some additional
code size trimming may happen, though the core of FreeRTOS is already
fairly light.