Pull in a heavily stripped copy of CherryUSB v1.1.0-29-g26e2436
This is intended to be just the files we need for a basic host with only
hub and msc support. Future updates pulled in from upstream should keep
that in mind. The idea is to limit excess files and duplication within
the repository from what's used by the normal firmware.