Initial consolidation of all USB serial interfaces
This is the first step towards having a common front-end for a wide
variety of USB serial drivers that all have compatibile interfaces and
almost-identical functionality. It eliminates a lot of needless
switch-case code on the integration side, and provides a point where we
can pull out more common code and add common functionality.