Major rework of FT260 meter probe USB driver
This is an on-going project, but a good time to make a checkpoint with
current progress.

Redesigned the FT260 HID front-end to use a similar task approach to the
other USB front-end control tasks.

Created a secondary "device" interface to the FT260 code that is more
persistent, tracking a particular device type and living beyond the
lifetime of a single USB device attach/detach. This makes it safer for
other code, like the meter probe task, to "hold onto" it.

Updated all relevant code to work with this new approach.

Added query to get the device serial number from the USB descriptor.