Update USB serial and keyboard task URB resubmit
These had originally been implemented using a timer, but it seems like
simply resubmitting the polling URBs on the handler task instead of the
ISR may be good enough. So these have been changed to match the way the
FT260 task handles it.

Timers for this are excessive and best avoided if unnecessary.