Add FT260 GPIO control and more TSL2585 VSYNC commands
These are functions of general utility leftover from the latest round of
VSYNC experiments.

The idea was to try and use a VSYNC pulse to trigger a measurement cycle
with known timing. While this did work, it unfortunately incurred the
same time delay overhead as earlier "single shot" experiments. This
makes it useless for the intended use case.