Initial implementation of "fast mode" for light sensor readings
The problem we face with the new USB<->I2C interface is that we can no
longer read the light sensor with the same level of responsiveness as
before.  Interrupts are captured within a 1ms window, and register reads
take a minimum of 3ms. Furthermore, quirks of the FT260 cause delays to
be inserted between byte reads that prevent I2C reads from running as
quickly as they should.

For normal sensor operations, none of this is a real problem. However,
enlarger calibration requires us to poll light readings at a much faster
rate. The solution to these problems, is to implement a dedicated "fast
mode" to change the way we read the sensor in these situations.

Fast mode works by doing the following:
- Runs FT260 I2C at 1MHz to lessen the impact of some of its timing
  oddities
- Leverages the sensor FIFO to read updates in groups of 8 readings,
  allowing for a 2.5ms integration time within a 20ms interrupt rate.
- Tries to work around any unexpected issues that occur when doing this,
  such as non-enabled sensor interrupts being triggered or spurious
  interrupt reports

The enlarger calibration code still needs to be rewritten to deal with
this new method of sensor reading, however it should still be a
straightforward process.  That code needs to know exactly when light
levels pass certain thresholds, but does not need to react to the change
very quickly.