Fix init and basic report handling
As long as we're going to assume that keyboards are handled according to
the boot protocol, the spec says we need to explicitly set the keyboard
to boot protocol mode during initialization.

Some devices may also send reports with different lengths than we
expect. For reports longer than 8 bytes, the spec days that the first
byte is the report ID, and if the device is in boot protocol mode, then
the next 8 bytes should still always match the standard report format.