Enable SOF interrupt and add various monitoring hooks
These are some minor changes to the CherryUSB library to help with
implementation of various error recovery methods.

Enabled the start-of-frame (SOF) interrupt in the dwc2 port, and added
a callback function to be notified of it. Not sure if the placement,
within the USB ISR is ideal, but the current plan is to simply use it to
increment a local timer. As such, placement does not matter much.

Added attach/detach callbacks for the hub class, similar to those that
exist for other USB class drivers. This is to make it possible to for
error recovery code to handle state changes in hub devices.