Implement watchdog on USB SOF and meter probe reattach
There are cases where a device glitch can cause the internal USB hub to
freeze up, in ways that it no longer detects connection events or causes
interrupts to be sent to the USB host stack. While the root cause of
these issues is hardware related, to have a robust system it needs to be
possible to recover from this in firmware.

At the most basic level, this involves being able to detect when the USB
peripheral stops sending start-of-frame (SOF) interrupts because it is
no longer communication with the internal hub IC, and to then toggle the
VBUS_DETECT line to cause the hub IC to reset its state.

In the USB host handler code, this is implemented via a watchdog timer.
This watchdog timer checks to see if the USB hub IC has connected within
the expected time period, and whether USB SOF interrupts are being
regularly fired once it is connected. If either case fails, then it
toggles the hub's VBUS_DETECT signal to reset its state.

On the meter probe side, a new "detached" state has been added. If the
meter probe task is running when this state is entered, then it will
automatically restart once the device is reattached. This may still need
more work to handle the unlikely case of two meter probes being
connected, even if to refuse to directly support it, but that's TBD.

There are also two more (related) failure modes that need to be handled,
albeit a bit differently. One is the case of a simple device failure
where the FT260 task falls out of one of its polling transactions. The
other is the same glitch that led us to this issue, but when there is an
intermediary hub. In both cases, the recovery may involve attempts at
resetting device communication or exercising some sort of direct port
control through the USB hub control interface.