Rearrange support components around the main board encoder
Moved the TVS diode protecting the encoder pins away from the underside
of the encoder, to a spot to the side and with a good path to the
mounting hole.

Rearranged the filter passives adjacent to the encoder for a cleaner
layout.

These changes cause the encoder signals to take a slightly more
roundabout path, but address some manufacturing concerns.

Fix #57