Increase trace clearance around display mounting nuts
Avoid getting traces too close to the display module mounting nuts, with
special attention to the side of the board those nuts are installed on.

This undoes some changes done for #47, however that was an attempt at
solving an issue more substantially affected by simply changing the slew
rate of the SPI GPIOs.

Fix #58