First attempt at updating USB connector footprint
This adds some paste areas around the mechanical pins, with hope that
it'll cause a bit of solder to wick onto the pins during reflow. This
isn't intended to solve the mechanical pin soldering problem, just to
keep the connectors in place before they can be manually touched up.

The first approach here is based on some suggestions seen in a
manufacturer app note, and involve shapes that KiCad doesn't make very
easy to implement.

Ref #61