Implement basic retry for USB control transfers
This replaces a number of hard-coded delays in the USB device
enumeration code with a basic retry loop for control transfers.
Some USB devices may fail with a TXERR during enumeration if requests
are made of them too quickly, and a retry solution for this was
noticed in another USB stack. This solution is a little bit further up
the stack than that one, but is simpler to implement and gets the job
done.