Initial implementation of live metering for B&W printing mode
When we're in B&W printing mode, and at least one meter reading has been
captured to anchor the tone scale, enable the live metering feature.

Live metering shows an overlay dot on top of the tone scale, showing the
latest reading at the spot over the meter probe. These readings are
quicker and less accurate than a normal meter reading, but give a good
feel for what the projected image is like in a particular spot.

Its still undecided whether this will be a configurable feature, or if
there will be some way for it to "go away" if the meter probe isn't
under the image (such as not displaying live readings far beyond one end of
the scale)