Change display footprint to use reverse entry connector
On the schematic side, the display is now split into two component
symbols. One for the display module itself, and one for the connector it
plugs into.

On the PCB layout side, the display module's footprint now only
represents the NPTH mounting holes, a variety of reference marks, and
the CAD model for the display itself.

The connector footprint is now separate, and uses a Harwin M20-7810845
reverse entry connector.  This should dramatically reduce the amount of
pin soldering necessary to install a display module, and make it easier
to remove a display module after installation.

Thus far, the mechanical mounting is unchanged pending additional
research into the various options.

In the process of making these changes, a lot of traces around the
display have been shoved around due to the fact that the new connector
isn't thru-hole on the connectivity side.

Refs #76