Add all design assets for the DensiStick peripheral
The Printalyzer DensiStick is a simple reflection-only densitometer
peripheral designed around the same basic components as the meter probe
and other internal projects.

It is designed to be driven entirely by a host computer/device, with
support to be added to both the Printalyzer Timer firmware and the
Densitometer Desktop software. Like the meter probe, its design is based
around a USB<->I2C bridge IC with some related GPIOs, and contains no
local logic of its own.

It is entirely up to the host side to control the sensor and LED driver,
and do the necessary calculations to produce density measurements.
However, it does contain a local EEPROM which stores both the USB device
properties and all of the relevant calibration data.

The current version of the design has been fully constructed into a
functional prototype, though support has not yet been added to the timer
firmware.

The basic goal of this device is to provide a low-cost way for users to
measure paper-profile test exposures created by contact-printing step
wedges like the Stouffer T2115. As such, the optical design is optimized
for narrow spot measurement and target alignment in one axis.

Fix #10