Update display footprint to use snap-in standoffs
This updates the CAD model assignments and positions for the display
module mounting footprint to use Keystone 8879 snap-in standoffs instead
of spacers for screws and nuts.  It also raises the display height from
3mm off the main PCB to 3.2mm.

This is going to be the new preferred way of mounting the display.  If
it doesn't work out, this change fortunately involves no actual PCB
design changes, so it will be easy to switch back at assembly time.

Ref #76