Move most main board components onto a metric grid
When this board was initially designed, everything (except connectors) was
done on a 5 mil grid. Over time, new areas of the board were instead laid
out on a metric (0.25mm or close) grid.

This now puts the board in a weird hybrid state, where components are
laid out on both types of grids, traces are generally in mils, and most
long traces are on a 5 mil grid.

As part of a cleanup, this commit hopefully puts all the components
(except the MCU section) onto a metric grid.  The MCU section should
probably be cleaned up as well, but it would be more difficult.
Likewise, most traces are still on that 5 mil grid due to difficulty of
cleaning up and tracking down all the stragglers.