Add initial panelization and automation bits
This brings in some KiKit infrastructure bits (tab markers, JSON config
files, support scripts) and a Makefile to run it all.

The actual panelization approach for the various boards in this project
is still TBD, pending conversation with assembly houses. However, this
puts in place some basic things to support it.

Ref #59