Various bits of TVS and layout cleanup for the meter probe and DensiStick
Decided to standardize on the new TI ESD321 TVS diode for single-line
protection on these boards. This part appears designed for this use,
especially with a lower working voltage. Its also in a similar package
to the Semtech part previously being used in some of these cases.

Created a new SOD-523 footprint based on TI's recommended land pattern,
which might work a little better than the stock one.

Updated the DensiStick to use the same 4.7uF MLCC for the bulk input
capacitor as the meter probe.

Various other minor tweaks to the fab layers and traces associated with
the above changes.