Fix USB FIFO size configuration
The CherryUSB library makes sure that the value of this define does not
exceed a value stored in the upper 16 bits of the GHWCFG3 register.

This register is not documented in the STM32F446 reference manual, but
it was found in this other document:
https://www.intel.com/content/www/us/en/programmable/hps/agilex5/topics/register_intreg_GHWCFG3.html

This says its the "DFIFO Depth", which would match with how its being
used.

The problem is that the value of this register section is slightly
smaller than what the default definitions of these macros assumed it to
be, causing the initialization checks to fail.

Whether or not this is even the right way to validate this value is
unknown, due to lack of clear documentation, but this quick fix seems to
work.