Remove all bootloader text in favor of graphics
This takes out all of the static text display screens, and thus avoids
the need to compile in U8G2 fonts or string handling code.
In its place, all the bootloader screens are replaced with various
graphics.

Initially, the space savings is mostly a wash, but that will change as
asset compression is added and more string code is stripped out.