Strip out code for unused peripherals
This removes the code to handle three timer channels, one SPI interface,
and sets a lot of I/O pins to an uninitialied/analog state.

Everything removed is either unnecessary for the bootloader, changed in
Rev F hardware, or defaults to a desired disabled state if a pin isn't
explicitly set.