Initial implementation of new firmware update process
There is now a menu in the main firmware's settings area where the user
can select an update to install. It will then pass this selection along
to the bootloader and restart the device to install it.

On the bootloader side, there are now three entry paths:
- The old path, button and known filename, for recovery
- The firmware-triggered path, for normal updates
- The checksum fail path, for recovering from missing or corrupted
  firmware in flash

The firmware passes the selection to the bootloader using an RTC backup
register as a flag, and a reserved block in the EEPROM for the actual
data.  The data consists of the serial number of the USB storage device,
the file path, and the file checksum.

This code still needs a lot of testing around its various paths, but
this is a good point to make a checkpoint commit.