Update firmware for new LED driver
Remove code for ST16CPC26 and add initial code for the AW9523

The way these two ICs operate is dramatically different, leading to
needing to make changes to the illumination controller and LED
diagnostic screen. Both of these may receive further changes in the
future.

With the ST16CPC26, on/off control and brightness are completely
separate from each other. Furthermore, brightness is global to all LEDs.

With the AW9523, each LED has its own brightness value which is also
used for on/off control. The plan is to take advantage of this at some
point in the future.