Change flash partition between bootloader and firmware
This change reduces the bootloader size from 128K to 64K,
and increases the application size from 384K to 448K.

The specific increments were chosen based on flash sector allocation of
the STM32F446, making it so that the first four 16K sectors go to the
bootloader and the remaining four sectors (64K, 3*128K) go to the
application code.

With this change, the bootloader allocation is now almost completely
full. However, any additional planned bootloader changes likely amount
to minor adjustments of control logic with little to no code size change.
If more space is needed for bootloader code, there are a number of
places it can be tightened up further without a lot of work. These
places include reducing/removing more logging code, refactoring to
reduce code duplication, and refactoring the code used to draw UI
graphics.

One noticed quirk is that the total size reported by the "size" command
on the initial ELF file generated by the build process is slightly
different from the actual flash size.  Running the "size" command on
processed output of the various "objcopy" steps returns the expected
values.

This does not seem to cause any problems, and appears to be due to minor
differences in the size of the padding around the "start_fill" and
"text" sections.