Remove meter probe sensor osc callibration code
Doing oscillator calibration requires a precise timing reference, which
is no longer practical to provide from the host over the new interface.

It can be provided via a crystal reference inside the meter probe
itself, but right now it doesn't seem to make enough of a difference to
the data to be necessary.

Stripping out this code as it is unused. Can re-add later in another
form if that changes.