Get the DensiStick task attaching to the device
This is the start of actually adding DensiStick support. It includes
making sure control tasks can attach to both the Meter Probe and the
DensiStick, and that we have diagnostic screens that can talk to either
device.

There is still a lot of work left to do, on the diagnostic screen, the
settings code, and the control interface.