Integrate densitometer measurements into the home display
The home display has a "Densitometer" mode, which was originally
conceived as a way of using the meter probe to capture relative density
measurements from an image projected onto the enlarger baseboard.

This commit does a major refactoring of that mode, to move around some
of its state data and incorporate the ability to also act as a display
for normal densitometers (DensiStick and USB/Serial attached).