Route VSYNC to FT260 GPIOA on DensiStick
This makes the sensor-to-bridge connection on the DensiStick logically
match the same from the Meter Probe, and makes it possible to experiment
with using VSYNC as a trigger.

Doing this may be useful to provide more ways of operating the sensor,
and doesn't hurt if not used.

More routing cleanup may be necessary before this design is considered
complete.