Handle repeated pipe stalls with a port reset
It is possible for the device to get wedged with a port stall retry loop
in some EFT burst situations. This handles a permanently stalled pipe by
resetting the hub port.

To make this work, the hub code needed to handle a port reset event by
triggering a device reconnect.

There are still cases where we won't fully recover, such as if errors
happen during the device re-enumeration process, but a user
unplug/replug cycle should fix that.

Ref #64