Update handling of slope correction to allow omitted data
The current calibration processes for the Meter Probe and the DensiStick
do not do slope correction. However, we want to keep the possibility of
having it within the calibration data and processing code. As such, the
code now checks if the values have been set, and can handle both cases.