Fix handling of blackout on startup
Because of how startup operations were ordered, there was a brief
flicker of panel illumination on startup before the blackout switch
setting would take effect.  This is a bad thing, because it should be
safe to restart the device under blackout conditions.

To fix it, a number of changes were made:
- Keypad controller initialization, including reading the initial state
  of the keypad, was moved outside of the keypad task and into the
  keypad init step
- The illumination controller now reads the initial state of the keypad
  before setting up its usual callback interactions.
- Some minor tweaks to keypad interrupt initialization to ensure that
  keypad events occuring between initalization and normal task operation
  are not missed

Fix #55